+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+- Added Android and iOS time change events.
+- Added a time event page to the example app.
+
 ## 0.5.0
 
 - Added Android and iOS orientation events.